# docs/swagger.yaml
definitions:
...
  consts.RoleStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      RoleStatusEnable: 启用
    x-enum-varnames:
    - RoleStatusEnable
  consts.UserGender:
    enum:
    - 0
    type: integer
    x-enum-comments:
      UserGenderNone: 未知
    x-enum-varnames:
    - UserGenderNone
  consts.UserStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      UserStatusEnable: 启用
    x-enum-varnames:
    - UserStatusEnable
...
  dto.ResetPwdDto:
    properties:
      password:
        description: 密码
        type: string
      userId:
        description: 用户Id
        type: integer
    type: object
  dto.SetUserStateDto:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
      userId:
        description: 用户Id
        type: integer
    type: object
...
  dto.SysRoleAddDto:
    properties:
      menuIds:
        description: 角色菜单
        items:
          type: integer
        type: array
      remark:
        description: 备注
        type: string
      roleCode:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.RoleStatus'
        description: 菜单状态（0正常 1停用）
    type: object
  dto.SysRoleUpdateDto:
    properties:
      id:
        description: 编号
        type: integer
      menuIds:
        description: 角色菜单
        items:
          type: integer
        type: array
      remark:
        description: 备注
        type: string
      roleCode:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.RoleStatus'
        description: 菜单状态（0正常 1停用）
    type: object
  dto.SysRoleVo:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 编号
        type: integer
      menuIds:
        description: 角色菜单
        items:
          type: integer
        type: array
      remark:
        description: 备注
        type: string
      roleCode:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.RoleStatus'
        description: 菜单状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
  dto.SysUserAddDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/consts.UserGender'
        description: 用户性别（0未知，1男，2女）
      nickName:
        description: 用户昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
      userName:
        description: 用户账号
        type: string
    type: object
  dto.SysUserUpdateDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/consts.UserGender'
        description: 用户性别（0未知，1男，2女）
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      status:
        allOf:
        - $ref: '#/definitions/consts.UserStatus'
        description: 帐号状态（0正常 1停用）
    type: object
  dto.SysUserVo:
    properties:
      avatar:
        description: 头像地址
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/consts.UserGender'
        description: 用户性别（0未知，1男，2女）
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色
        items:
          type: integer
        type: array
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户账号
        type: string
      userType:
        description: 用户类型（0普通账号，1超级管理员）
        type: integer
    type: object
...
  request.IdInfoDto:
    properties:
      id:
        description: id编号
        type: integer
    type: object
  response.JsonResult:
    properties:
      code:
        description: 状态码（架构层面：中间件/异常处理等）
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      success:
        description: 是否成功（业务层面）
        type: boolean
    type: object
  response.PageResult:
    properties:
      list:
        description: 数据
      totalCount:
        description: 总条数
        type: integer
    type: object
info:
  contact: {}
  description: go-fast-admin 接口文档
  title: 系统模块
  version: 1.0.0
paths:
...
  /system/role/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加角色
      tags:
      - 角色
  /system/role/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - 角色
  /system/role/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysRoleVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取角色详情
      tags:
      - 角色
  /system/role/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysRoleVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 角色列表
      tags:
      - 角色
  /system/role/query:
    get:
      parameters:
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 角色名称
        in: query
        name: roleName
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysRoleVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 角色查询
      tags:
      - 角色
  /system/role/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysRoleUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - 角色
  /system/user/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 用户
  /system/user/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
  /system/user/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysUserVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 用户
...
  /system/user/query:
    get:
      parameters:
      - description: 用户昵称
        in: query
        name: nickName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户查询
      tags:
      - 用户
  /system/user/resetPwd:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPwdDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 重置密码
      tags:
      - 用户
  /system/user/setStatus:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SetUserStateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 设置用户状态
      tags:
      - 用户
  /system/user/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"